Update fs.js
In new versions of Node (tested in v. 0.11.13), `Function.name` is a read-only property [(it is likely to stay this way)](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Function/name), assigning a value to it causes an error in strict mode.
I presume that this is only for debugging reasons, so I left the catch clause empty.